+++ src/Exceptions/PayPalDetailsMissingOnOrderException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace DoubleThreeDigital\SimpleCommerce\Exceptions;
+
+class PayPalDetailsMissingOnOrderException extends \Exception
+{
+    //
+}