+++ src/Http/Controllers/GatewayCallbackController.php
@@ -16,7 +16,8 @@ class GatewayCallbackController extends BaseActionController
 
     public function index(Request $request, $gateway)
     {
-        $order = Order::find($request->get('_order_id'));
+        // $order = Order::find($request->get('_order_id'));
+        $order = $this->getCart();
 
         $gatewayName = $gateway;